[plugin] ConvDiff: adding 2nd source term ("surface source")
Adding ability to ConvectionDiffusion plugin to cope with a second
source term ("surface source") as it occures in

\nabla\cdot ( -D\nabla\vec{c} + \vec{v} ) = 0

Usage:
In line with current behaviour the surface source term is only used if
manually set via elemDisc.set_surface_source(). Similar to the velocity,
it expects either a vector or the vector's elements separately.

See apps/conv_diff/laplace_fe.lua for an example.

Implementation details:
This commit introduces an additional member variable to store the
surface source and the corresponding setter functions (set_surface_source)
in analogy to the ones for the velocity. Also the implementation of the
FE discretization has been adjusted and extended to incorporate this new
term.

Note:
Utilization of this surface source is not yet implemented for FV or
other discretization methods.